Improve code generated by `kernpairs` statement
In addition to ordering the generated rules by UFO-specified priority (e.g. glyph-to-glyph first, glyph-to-class next, etc), this code now converts glyph-to-class rules to glyph-to-glyph rules where possible (a) to make the GPOS smaller and, more importantly, (b) to resolve some rule conflicts that otherwise would cause generated fea code to be unreachable in the GPOS table.