draft of implementation of `--compregex` option to remove unreachable component glyphs
Added `--compregex` to enable a subsetter pass that searches for and removes unreachable component-only glyphs. The option's argument is to be a regex; all glyphs whose names match the regex will be assumed to be component-only glyphs that can be considered for removal. If `--compregex` is followed immediately by another commandline option, then `^_` is assumed for the regex, thus all glyphs starting with underscore will be considered for removal.